Use set boundaries for the 2d grid
Now the dimensions of the 2d grid used for computation are static and
know beforehand, so the from-pos and to-pos arguments are not needed
anymore.

They can be replaced with a single
(set-dimensions (cons from-pos to-pos))
before the automata are intialized.

This change allows for more accurate neighborhoods, since now the
neighborhood “wraps around” the boundaries of the grid, so every
automaton has the same number of neigbors. Due to this change, e. g.
Conway's Game of Life can be implemented correctly.